Added an /add?url method so URLs can be added through the web server.
Pass the server particulars (address, port, name) into server sessions
so they can be included in messages.

Build a bookmarklet (javascript bookmark) on the main page from the
address, port, and name.  The bookmarklet can be bookmarked on the
toolbar and will submit whatever page is referenced by the
location/address field.